+++ src/kernel/include/filesystem/fscalls.h
@@ -34,7 +34,7 @@ int open(const char *pathname, int flags);
 int close(int fd);
 int lseek(int fd, off_t offset, int whence);
 int getdents(int fd, dirent *dirp, unsigned int count);
-int getcwd(char *buf, int size);
+int getcwd(char *buf, size_t size);
 int chdir(const char *path);
 int read(int fd, void *buf, size_t count);
 int write(int fd, const void *buf, size_t count);
